Remove connect() timeout delay
Let the targeted BGP daemon handle the simultaneous connections
pressure.

Tip: to decrease the convergence time caused by the connections timeout
or reject, please increase the tcp socket listen connection queue.

Example:
  listen(socket, 3);
to
  listen(socket, SOMAXCONN);

For more information read the listen() man page.

Test notes: it was noticed that using the SOMAXCONN was already enough
to handle a thousand simultaneous peers connecting, but if you want to
handle even more simultaneous connections you have to also bump the
sysctl value:
net.core.somaxconn

The default value is set to SOMAXCONN which is 128 as of this writting.